pyproject.toml
[project]
name = "hello-foundation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "sc-foundation-services>=3.0.3",
]
